Set map height based on parent size
fix glitch
